test: Refactor and fix event listening code
Event listening code was abstracted into a `EventListenerService` class.
Now we don't have to repeat the creation of listener+promise, and remove
the listener manually.

Also, using this class we were able to fix the issue of dangling event
listeners, preventing the test process from exiting (open handles). We
fix it by calling `reset()` on `EventListenerService` in the `afterAll`
hook of the tests, which will do this for us, without any errors.